Feat: type alias 들을 types/History.ts 파일로 뺌 (#161)
dependency cycle detected 에러 때문에 History의 data type들을 외부에서 따로 관리함
