+++ client/src/components/HistoryList/HistoryItem.tsx
@@ -1,20 +1,21 @@
 import Icon from '../common/Icon'
 import styles from './HistoryItem.module.scss'
 import { timerFormat } from '../../utils/date'
-import { ItemType } from './Calendar/Calendar'
+import { WalkLogContentsDataType } from '../../types/History'
+import { differenceInSeconds } from '../../utils/date-fns'
 
 type HistoryItemProps = {
-  item: ItemType
+  item: WalkLogContentsDataType
   startAt: string
 }
 
 export default function HistoryItem({ item, startAt }: HistoryItemProps) {
-  const { id, imageUrl, text, createdAt } = item
-  const snapTimeDiff = new Date(createdAt).getTime() - new Date(startAt).getTime()
-  const snapTime = timerFormat(snapTimeDiff / 1000)
+  const { walkLogContentId, imageUrl, text, createdAt } = item
+  const snapTimeDiff = differenceInSeconds(new Date(createdAt), new Date(startAt))
+  const snapTime = timerFormat(snapTimeDiff)
 
   return (
-    <li key={id} className={styles.container}>
+    <li key={walkLogContentId} className={styles.container}>
       {imageUrl && <img src={imageUrl} alt='올린 사진' />}
       <div className={imageUrl ? styles.timeTextBox : styles.noImg}>
         <div className={styles.snapTimeBox}>
