--- client/src/types/HistoryDetail.ts
@@ -1,11 +1,0 @@
-export type WalkLogContentsDataType = {
-  createdAt: string
-  imageUrl: string | null
-  text: string | null
-  walkLogContentId: string
-}
-
-export type ModalOption = {
-  title: string
-  deleteFn: () => void
-}
